# [stdlib] Add benchmarks for Character's unicodeScalars view