@dabrahams 2017-06-28
@swift-ci Please smoke benchmark