@palimondo 2017-07-06
Can we do something to the naming of these tests? They don't follow the convention established so far… and also are horribly long.