@palimondo 2017-07-06
I suggest:
* drop underscores
* use abbreviated country codes and proper capitalization of abbreviations
* drop `unicodeScalars` suffix - they all have it…
* shorten Backwards to Back

E.g.: `CharIterationRU`, `CharIterationBackRU`, `CharIterationUTF16`, etc.