@palimondo 2017-07-06
Or `CharIterationRevCN`, as in reverse (instead of `Back`)