@palimondo 2018-12-12
@dabrahams I'm looking over these benchmarks and wonder about the motivation for testing `.reversed` iteration and indexing. I can see this might be important in character iteration for Boyer–Moore string-search algorithm. But why is this applicable on the `unicodeScalar` level?