@dabrahams 2018-12-12
I'm sorry, I don't think I can offer any real insight.  I think the idea was to make sure all the new components were performant and didn't regress.