@palimondo 2018-12-12
Also, I have tried to test locally a modified versions without first copying the `Character`s to `Array`, thinking `String` is fully materialized collection, there shouldn't be much difference, only to find 3-4x slowdown (30-60 on "utf16" variant), both for imperative `for-each` approach and functional `map.reduce`. I was hoping to recover some performance by going `.lazy`, but it didn't help.