@palimondo 2018-12-12
How realistic is testing iteration over `Character`'s `unicodeScalar` view from `Array`, and not directly from `String`?