@milseman 2018-12-17
> But why is this applicable on the unicodeScalar level?

Reversed decoding can be more expensive, and is not optimizable by using a customized iterator.

> How realistic is testing iteration over Character's unicodeScalar view from Array, and not directly from String?

Sorry, could you re-phrase this? BTW, I introduced a custom iterator for String.UnicodeScalarView, which makes pure iteration much faster than the indexing approach.