@palimondo 2018-12-17
I was questioning the practical applicability (or realism) of how this benchmark is currently built. I can understand the importance of reverse iteration on the character view as that's used for example in the Boyer–Moore string-search algorithm. I guess this is also applicable to the `String`'s `.unicodeScalar` view… I see "backwards iteration" mentioned in the TR on [Unicode text segmentation](https://unicode.org/reports/tr29/#Random_Access) boundary detection and such. 

But the `CharIteration` and `CharIndexing` benchmarks seem too synthetic to me,  because they focus only on testing **`Character`**'s `.unicodeScalar` view. Of all the workloads, only the `utf16` contains different number of characters vs. unicodeScalars. Which means we are mostly testing forward and `.reverse` iteration on a single element collections retrieved from `Character`. Moreover, these `Character`s are not accessed from a`String`, as I guess would be typical in string processing, but are precomputed into `Array<Character>`. So I think we end up redundantly measuring a really thin aspect here.

Shouldn't these benchmarks be changed to measure forward and reverse iteration on the `characters` and `unicodeScalar` views from the String? As mentioned above, that results in runtimes that are 3-6x slower (30-60 on "utf16" variant) than the synthetic-once we currently have. It looks like there is quite significant overhead in creating the individual `Character`'s from the strings?